Fix types

+++ nes/bus.py
@@ -1,8 +1,5 @@
 from typing import List
-
 import numpy as np
-
-
 
 class Bus:
     """
@@ -14,16 +11,16 @@ class Bus:
         
         # Devices on the bus:
         self.cpu : CPU = CPU()
-        self.ram : List[np.uint8] = [0x00] * 1024 * 64 # fake RAM for now
+        self.ram : List[int] = [0x00] * 1024 * 64 # fake RAM for now
         
         # Connect CPU to the bus:
         self.cpu.connect_bus(self)
     
-    def write(self, address: np.uint16, data: np.uint8) -> None:
+    def write(self, address: int, data: int) -> None:
         if 0x0000 <= address <= 0xFFFF:
             self.ram[address] = data
         
-    def read(self, address: np.uint16, read_only: bool = False) -> np.uint8:
+    def read(self, address: int, read_only: bool = False) -> int:
         if 0x0000 <= address <= 0xFFFF:
             return self.ram[address]
         return 0x00
